+++ app/models/maths_and_physics/slug_sequence.rb
@@ -25,6 +25,7 @@ module MathsAndPhysics
       "information-provided",
       "personal-details",
       "postcode-search",
+      "no-address-found",
       "select-home-address",
       "address",
       "gender",