+++ app/models/student_loans/slug_sequence.rb
@@ -21,6 +21,7 @@ module StudentLoans
       "information-provided",
       "personal-details",
       "postcode-search",
+      "no-address-found",
       "select-home-address",
       "address",
       "gender",